移除sign模块，重写至yiban/
