--- application/sign/Model/WkPageLinks.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace app\sign\model;
-
-use think\Model;
-use think\Request;
-
-/**
- * 统一授权模型
- * @package app\admin\model
- */
-class WkPageLinks extends Model{
-	protected $connection = 'mysql://root:69431589@127.0.0.1:3306/wiki#utf8';
-}
