--- application/sign/common.php
@@ -1,5 +1,0 @@
-<?php
-
-use think\Route;
-
-Route::rule('uiop','sign/api/index','GET');
