--- application/sign/view/index/show.html
@@ -1,59 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-    <!-- 引入 ECharts 文件 -->
-    <script src="//cdn.bootcss.com/jquery/3.1.1/jquery.min.js"></script>
-    <script src="//cdn.bootcss.com/echarts/3.4.0/echarts.js"></script>
-</head>
-<body>
-    <!-- 为ECharts准备一个具备大小（宽高）的Dom -->
-    <div id="main" style="width: 1000px;height:800px;"></div>
-    <script type="text/javascript">
-var myChart = echarts.init(document.getElementById('main'));
-myChart.showLoading();
-$.get('/dolphinphp/node.json', function (webkitDep) {
-    myChart.hideLoading();
-
-    option = {
-        legend: {
-            data: ['HTMLElement', 'WebGL', 'SVG', 'CSS', 'Other']
-        },
-        series: [{
-            type: 'graph',
-            layout: 'force',
-            legendHoverLink:true,
-            animation: false,
-            // symbol:'roundRect',
-            edgeSymbol: ['circle', 'arrow'],
-            edgeSymbolSize:4,
-            label: {
-                normal: {
-                    //show:'true',
-                    position: 'right',
-                    formatter: '{b}',
-                    textStyle:{
-                        fontSize:6,
-                    }
-                }
-            },
-            draggable: true,
-            data: webkitDep.nodes.map(function (node, idx) {
-                node.id = idx;
-                return node;
-            }),
-            categories: webkitDep.categories,
-            force: {
-                edgeLength: 10,
-                repulsion: 20,
-                gravity: 0.2
-            },
-            edges: webkitDep.links
-        }]
-    };
-
-    myChart.setOption(option);
-});
-    </script>
-</body>
-</html>
